# LocadoraAPI/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LocadoraAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Alugueis": {
      "post": {
        "tags": [
          "Alugueis"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AluguelCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AluguelCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AluguelCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Aluguel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Aluguel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Aluguel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Clientes/cpf/{cpf}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          }
        }
      }
    },
    "/api/Veiculos": {
      "get": {
        "tags": [
          "Veiculos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veiculo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veiculo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veiculo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Veiculos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          }
        }
      }
    },
    "/api/Veiculos/{id}": {
      "get": {
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Veiculos/placa/{placa}": {
      "get": {
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "name": "placa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Aluguel": {
        "required": [
          "clienteId",
          "dataFim",
          "dataInicio",
          "veiculoId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "veiculoId": {
            "type": "integer",
            "format": "int32"
          },
          "veiculo": {
            "$ref": "#/components/schemas/Veiculo"
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "dataInicio": {
            "type": "string",
            "format": "date-time"
          },
          "dataFim": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AluguelCreateDto": {
        "type": "object",
        "properties": {
          "veiculoId": {
            "type": "integer",
            "format": "int32"
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "dataInicio": {
            "type": "string",
            "format": "date-time"
          },
          "dataFim": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Cliente": {
        "required": [
          "cpf",
          "email",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "cpf": {
            "maxLength": 14,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "telefone": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClienteCreateDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClienteUpdateDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Veiculo": {
        "required": [
          "ano",
          "marca",
          "modelo",
          "placa",
          "valorDiaria"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "marca": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "modelo": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "placa": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "valorDiaria": {
            "type": "number",
            "format": "double"
          },
          "disponivel": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "VeiculoCreateDto": {
        "type": "object",
        "properties": {
          "marca": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "placa": {
            "type": "string",
            "nullable": true
          },
          "valorDiaria": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "VeiculoUpdateDto": {
        "type": "object",
        "properties": {
          "marca": {
            "type": "string",
            "nullable": true
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "valorDiaria": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Alugueis"
    },
    {
      "name": "Clientes"
    },
    {
      "name": "Veiculos"
    }
  ]
}